update routes for only notif type
